Move core task entities to feature module
